+++ src/docker/bin/docker-common.sh
@@ -12,12 +12,12 @@ DOCKER_LOGUSAGE=${DOCKER_LOGUSAGE-usage}
 # Write messages to console if interactive or syslog if not.
 # Usage: inform priority message
 # The priority may be specified numerically or as a facility.level pair.
-# Example user.notice, or 1.6 level is one of: 
+# Example user.notice, or 1.6 level is one of:
 # 0|emerg|1|alert|2|crit|3|err|4|warning|5|notice|6|info|7|debug
 #
 dc_log() {
 	local script=$(basename $0)
-	local stamp=$(dc_log_stamp)
+	local stamp="$(dc_log_stamp)"
 	local prio=$1
 	local level=${prio#*.}
 	local logtag="${script}[${$}]"
@@ -104,16 +104,25 @@ dc_update_loglevel() {
 #
 # Print package versions
 #
-dc_apk_versions() {
+dc_pkg_versions() {
 	local pkgs="$@"
 	local len=$(echo $pkgs | tr " " "\n" | wc -L)
-	local ver
+	local ver ver_cmd sed_flt
 	local os=$(sed -rn 's/PRETTY_NAME="(.*)"/\1/p' /etc/os-release)
 	local kern=$(uname -r)
 	local host=$(uname -n)
 	dc_log 5 $host $os $kern
+	if [ -x "$(command -v apk)" ]; then
+		ver_cmd="apk info -s"
+		sed_flt="s/.*-(.*)-.*/\1/p"
+	elif [ -x "$(command -v apt)" ]; then
+		ver_cmd="apt list --installed"
+		sed_flt="s/[^ ]+ ([[:digit:]]+\.[[:digit:]]+\.[[:digit:]]+).*/\1/p"
+	else
+		dc_log 5 "No package manager found among: apk apt"
+	fi
 	for pkg in $pkgs; do
-		ver=$(apk info -s $pkg 2> /dev/null | sed -rn 's/.*-(.*)-.*/\1/p')
+		ver=$($ver_cmd $pkg 2> /dev/null | sed -rn "$sed_flt")
 		if [ -n "$ver" ]; then
 			printf "\t%-${len}s\t%s\n" $pkg $ver
 		fi