+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [postfix](src/postfix) Use `smtpd_reject_unlisted_recipient=no` in `postfix_setup_domains()`.
 - [dovecot](src/dovecot) Remove `smtps/inet/smtpd_recipient_restrictions` and rely on `smtpd_relay_restrictions` instead in `dovecot_setup_smtpd_sasl()`.
 - [postfix](src/postfix) Allow recipient email address to be rewritten using regexp in `REGEX_ALIAS`.
+- [docker](src/docker/bin/docker-common.sh) Print package versions function `dc_pkg_versions()` now supports both apk and apt.
 
 # 1.4.2